HMAN-27- code refactor
