+++ src/main/java/uk/gov/hmcts/reform/hmc/helper/CaseCategoriesMapper.java
@@ -5,7 +5,6 @@ import uk.gov.hmcts.reform.hmc.data.CaseCategoriesEntity;
 import uk.gov.hmcts.reform.hmc.data.CaseHearingRequestEntity;
 import uk.gov.hmcts.reform.hmc.model.CaseCategory;
 import uk.gov.hmcts.reform.hmc.model.CaseCategoryType;
-import uk.gov.hmcts.reform.hmc.model.HearingRequest;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,10 +15,9 @@ public class CaseCategoriesMapper {
     public CaseCategoriesMapper() {
     }
 
-    public List<CaseCategoriesEntity> modelToEntity(HearingRequest hearingRequest,
+    public List<CaseCategoriesEntity> modelToEntity(List<CaseCategory> caseCategories,
                                                     CaseHearingRequestEntity caseHearingRequestEntity) {
         List<CaseCategoriesEntity> caseCategoriesEntities = new ArrayList<>();
-        List<CaseCategory> caseCategories = hearingRequest.getCaseDetails().getCaseCategories();
         for (CaseCategory category : caseCategories) {
             final CaseCategoriesEntity categoryEntity = new CaseCategoriesEntity();
             categoryEntity.setCategoryType(CaseCategoryType.valueOf(category.getCategoryType()));
