+++ src/main/java/uk/gov/hmcts/reform/hmc/helper/RequiredFacilitiesMapper.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.hmc.helper;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.hmc.data.CaseHearingRequestEntity;
 import uk.gov.hmcts.reform.hmc.data.RequiredFacilitiesEntity;
-import uk.gov.hmcts.reform.hmc.model.HearingRequest;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,10 +13,9 @@ public class RequiredFacilitiesMapper {
     public RequiredFacilitiesMapper() {
     }
 
-    public List<RequiredFacilitiesEntity> modelToEntity(HearingRequest hearingRequest,
+    public List<RequiredFacilitiesEntity> modelToEntity(List<String> facilities,
                                                         CaseHearingRequestEntity caseHearingRequestEntity) {
         List<RequiredFacilitiesEntity> requiredFacilitiesEntities = new ArrayList<>();
-        List<String> facilities = hearingRequest.getHearingDetails().getFacilitiesRequired();
         for (String facility : facilities) {
             final RequiredFacilitiesEntity requiredFacilitiesEntity = new RequiredFacilitiesEntity();
             requiredFacilitiesEntity.setFacilityType(facility);
