+++ src/test/java/uk/gov/hmcts/reform/hmc/service/HearingManagementServiceTest.java
@@ -15,6 +15,7 @@ import uk.gov.hmcts.reform.hmc.utils.TestingUtil;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -95,6 +96,10 @@ class HearingManagementServiceTest {
         hearingRequest.setHearingDetails(TestingUtil.hearingDetails());
         hearingRequest.getHearingDetails().setPanelRequirements(TestingUtil.panelRequirements());
         hearingRequest.setCaseDetails(TestingUtil.caseDetails());
+        HearingEntity hearingEntity = new HearingEntity();
+        given(hearingMapper.modelToEntity(hearingRequest)).willReturn(TestingUtil.hearingEntity());
+        given(hearingRepository.save(hearingEntity)).willReturn(TestingUtil.hearingEntity());
+        given(hearingRepository.findById(1L)).willReturn(Optional.of(TestingUtil.hearingEntity()));
         hearingManagementService.saveHearingRequest(hearingRequest);
     }
