+++ src/test/java/uk/gov/hmcts/reform/hmc/utils/TestingUtil.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.reform.hmc.utils;
 
 import org.assertj.core.util.Lists;
+import uk.gov.hmcts.reform.hmc.data.CaseHearingRequestEntity;
 import uk.gov.hmcts.reform.hmc.data.HearingEntity;
 import uk.gov.hmcts.reform.hmc.model.CaseCategory;
 import uk.gov.hmcts.reform.hmc.model.CaseDetails;
@@ -116,7 +117,7 @@ public class TestingUtil {
 
     public static HearingResponse hearingResponse() {
         HearingResponse response = new HearingResponse();
-        response.setHearingRequestId(2000000L);
+        response.setHearingRequestId(1L);
         response.setStatus(HEARING_STATUS);
         response.setTimeStamp(LocalDateTime.now());
         return response;
@@ -124,8 +125,9 @@ public class TestingUtil {
 
     public static HearingEntity hearingEntity() {
         HearingEntity hearingEntity = new HearingEntity();
-        hearingEntity.setId(2000000L);
+        hearingEntity.setId(1L);
         hearingEntity.setStatus(HEARING_STATUS);
+        hearingEntity.setCaseHearingRequest(new CaseHearingRequestEntity());
         return hearingEntity;
     }
